+++ include/simpledbus/advanced/Proxy.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <simpledbus/advanced/Callback.h>
+#include <external/kvn_safe_callback.hpp>
 #include <simpledbus/advanced/Interface.h>
 
 #include <memory>
@@ -48,8 +48,8 @@ class Proxy {
     void message_forward(Message& msg);
 
     // ----- CALLBACKS -----
-    Callback<std::function<void(std::string)>, std::string> on_child_created;
-    Callback<std::function<void(std::string)>, std::string> on_child_signal_received;
+    kvn::safe_callback<void(std::string)> on_child_created;
+    kvn::safe_callback<void(std::string)> on_child_signal_received;
 
     // ----- TEMPLATE METHODS -----
     template <typename T>